requirements.txt
streamlit
requests
flask
flask-cors
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai